Fix fixed stake reward check
The stake reward, even if fixed, includes the fees of the transactions included in the block.

The removed check does not take the fees making the network reject all the blocks which include a transaction. At this point, the total amount of fees is still unknown.

This is safely removed as the check is properly done in line 3322.