Fix to staking parameter "-mininputvalue"
The nMinimumInputValue variable controls the minimum value
that can be selected for staking.
However, when the variable is defined, the parameter "-mininputvalue"
is not yet available. This makes the parameter irrelevant.

To correct this, the value of the variable is refreshed later, just
before it is actually used.

The correction does not seem to have any side effects, apart from
the improper use of the parameter itself.

Resolve: #585