Patch UpdatePrice if curl don't get data on WIN
In some cases the UpdatePrice function does not update the data
because curl cannot establish a secure connection.

With this fix, if curl has reported error code 60, attempts a new
connection without peer verification, like DownloadBlockchain() do.

If the remedy works, this shortcut will be used for the rest of the
session to recover the prices.
Otherwise it will not try the remedy again.